Reduce size of shard info and turn it into a struct (#1984)
This replaces the `[u64; 2]` shard info with `ShardInfo { id: u32, total: u32 }`.

+++ src/client/bridge/voice/mod.rs
@@ -14,7 +14,7 @@ pub trait VoiceGatewayManager: Send + Sync {
     /// Performs initial setup at the start of a connection to Discord.
     ///
     /// This will only occur once, and provides the bot's ID and shard count.
-    async fn initialise(&self, shard_count: u64, user_id: UserId);
+    async fn initialise(&self, shard_count: u32, user_id: UserId);
 
     /// Handler fired in response to a [`Ready`] event.
     ///
@@ -22,7 +22,7 @@ pub trait VoiceGatewayManager: Send + Sync {
     /// once per active shard.
     ///
     /// [`Ready`]: crate::model::event::Event
-    async fn register_shard(&self, shard_id: u64, sender: Sender<InterMessage>);
+    async fn register_shard(&self, shard_id: u32, sender: Sender<InterMessage>);
 
     /// Handler fired in response to a disconnect, reconnection, or rebalance.
     ///
@@ -30,7 +30,7 @@ pub trait VoiceGatewayManager: Send + Sync {
     /// Unless the bot is fully disconnecting, this is often followed by a call
     /// to [`Self::register_shard`]. Users may wish to buffer manually any gateway messages
     /// sent between these calls.
-    async fn deregister_shard(&self, shard_id: u64);
+    async fn deregister_shard(&self, shard_id: u32);
 
     /// Handler for VOICE_SERVER_UPDATE messages.
     ///
